scooper/requirements.txt
aws-cdk-lib==2.137
boto3~=1.34
click~=8.1
constructs>=10.0.0,<11.0.0
PyYAML~=6.0
tqdm~=4.66
